Remove RadioScheduleData component.
