+++ src/app/containers/RadioSchedule/helpers/validServices.js
