+++ src/app/containers/RadioSchedule/index.jsx
@@ -1,20 +1,54 @@
-import React, { useContext } from 'react';
-import RadioScheduleData from './RadioScheduleData';
-import { ServiceContext } from '#contexts/ServiceContext';
+import React, { useEffect, useState } from 'react';
+import 'isomorphic-fetch';
+import { string } from 'prop-types';
+import webLogger from '#lib/logger.web';
 
-const getLocalRadioScheduleEndpoint = service => {
-  const localhostURL = 'http://localhost:7080';
-  const localServiceURL = `${localhostURL}/${service}`;
+const logger = webLogger();
 
-  return `${localServiceURL}/bbc_${service}_radio/most_read.json`;
-};
+const RadioScheduleContainer = ({ endpoint }) => {
+  const [schedule, setRadioSchedule] = useState([]);
 
-const RadioScheduleContainer = () => {
-  const { service } = useContext(ServiceContext);
+  const handleResponse = async response => {
+    const radioScheduleData = await response.json();
+    setRadioSchedule(radioScheduleData.schedules);
+  };
+
+  useEffect(() => {
+    const fetchRadioScheduleData = pathname =>
+      fetch(pathname, { mode: 'no-cors' })
+        .then(handleResponse)
+        .catch(e => logger.error(`HTTP Error: "${e}"`));
+
+    fetchRadioScheduleData(endpoint);
+  }, [endpoint]);
 
   return (
-    <RadioScheduleData endpoint={getLocalRadioScheduleEndpoint(service)} />
+    <>
+      <p>Radio Schedules</p>
+      {schedule.map(
+        ({
+          broadcast,
+          transmissionTimeStart,
+          transmissionTimeEnd,
+          episode: {
+            presentationTitle,
+            synopses: { short },
+          },
+        }) => (
+          <ul key={broadcast.pid}>
+            <li>{presentationTitle}</li>
+            <li>{short}</li>
+            <li>{transmissionTimeStart}</li>
+            <li>{transmissionTimeEnd}</li>
+          </ul>
+        ),
+      )}
+    </>
   );
+};
+
+RadioScheduleContainer.propTypes = {
+  endpoint: string.isRequired,
 };
 
 export default RadioScheduleContainer;
